Move GutenbergAttributeDecoder to separate file.
